+++ changelog.md
@@ -1,4 +1,4 @@
 # Changelog
 
-**./test-parts/backup**
-* Deleted file
+**./test-parts/test-possible-routes.js**
+* Uncommented all unit test calls.